* commit '54106b9ea073069219aa6d358985d005684eda02':
  bug fix HTC-1010

+++ src/main/resources/db/migration/postgres/V20200904001900__Damir_alter_application.sql
@@ -1,2 +1,2 @@
 alter table htc_dm_application
-  add column view_count numeric(19) null;
+  add column view_count integer null;
